created account details repository,service

+++ NetBankingProject/src/main/java/com/mindgate/main/domain/ChequeDetails.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class ChequeDetails {
 
-	private int chequeNo;
+	private long chequeNo;
 	private AccountDetails issuerAccNo;
 	private AccountDetails receiverAccNo;
 	private double chequeAmount;
